added 403 and 404 status code for DELETE theme

+++ src/theme/theme.controller.ts
@@ -7,15 +7,21 @@ import {
   Param,
   Delete,
   Query,
+  ForbiddenException, NotFoundException
 } from '@nestjs/common';
 import { ThemeService } from './theme.service';
+import { PostService } from '../post/post.service';
 import { CreateThemeDto } from './dto/createThemeDto';
 import { ThemeEntity } from './theme.entity';
 import { UpdateThemeDto } from './dto/updateThemeDto';
 
 @Controller('theme')
 export class ThemeController {
-  constructor(private themeService: ThemeService) {}
+  constructor(
+    private themeService: ThemeService,
+    private postService: PostService,
+  ) {}
+
   @Post()
   create(@Body() createThemeDto: CreateThemeDto) {
     return this.themeService.create(createThemeDto);
@@ -37,7 +43,17 @@ export class ThemeController {
   }
 
   @Delete(':id')
-  remove(@Param('id') id: number) {
-    return this.themeService.remove(id);
+  async remove(@Param('id') id: number) {
+    const theme = await this.themeService.find(id);
+    if (!theme) {
+      throw new NotFoundException('This theme does not exist');
+    }
+    const posts = await this.postService.findAll(id, null);
+    if (posts.length > 0) {
+      throw new ForbiddenException('This theme has posts linked to it');
+    }
+
+    await this.themeService.remove(id);
+    return { message: 'Theme deleted successfully' };
   }
 }
